# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dubcast Radio API",
    "description": "API интернет-радио Dubcast.\nСодержит публичные, пользовательские и админские эндпоинты.\n",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/playlists/import": {
      "post": {
        "tags": [
          "Playlists"
        ],
        "summary": "Import playlist from URL",
        "description": "Imports a playlist from an external URL (for example, SoundCloud playlist URL),\nparses its metadata and stores it as a new playlist in the system.\n",
        "operationId": "importPlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Playlist successfully imported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported URL."
          },
          "500": {
            "description": "Unexpected server error during import."
          }
        }
      }
    },
    "/api/parser/track": {
      "post": {
        "tags": [
          "Parser"
        ],
        "summary": "Parse a single SoundCloud track",
        "description": "Accepts a SoundCloud track URL and returns parsed track metadata.",
        "operationId": "parseTrack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Track was successfully parsed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL format or unsupported resource."
          },
          "500": {
            "description": "Unexpected server error while contacting SoundCloud or parsing the track."
          }
        }
      }
    },
    "/api/parser/playlist": {
      "post": {
        "tags": [
          "Parser"
        ],
        "summary": "Parse a SoundCloud playlist",
        "description": "Accepts a SoundCloud playlist URL and returns metadata for all tracks in the playlist.",
        "operationId": "parsePlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Playlist was successfully parsed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL format or unsupported resource."
          },
          "500": {
            "description": "Unexpected server error while contacting SoundCloud or parsing the playlist."
          }
        }
      }
    },
    "/api/parser/duration": {
      "post": {
        "tags": [
          "Parser"
        ],
        "summary": "Get track duration in seconds",
        "description": "Accepts a SoundCloud track URL and returns its duration in seconds.",
        "operationId": "getDuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Track not found or duration could not be determined.",
            "content": {

            }
          },
          "200": {
            "description": "Duration was successfully calculated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Duration of the track in seconds."
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL format or unsupported resource."
          },
          "500": {
            "description": "Unexpected server error while contacting SoundCloud or parsing metadata."
          }
        }
      }
    },
...
    "/api/playlists": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "summary": "Get all playlists",
        "description": "Returns a list of all playlists available in the system.",
        "operationId": "getAll_3",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "List of playlists successfully returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{id}": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "summary": "Get playlist by ID",
        "description": "Returns a single playlist by its identifier.",
        "operationId": "getById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the playlist",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Playlist with the given ID was not found.",
            "content": {

            }
          },
          "200": {
            "description": "Playlist found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Playlists"
        ],
        "summary": "Delete playlist",
        "description": "Deletes a playlist by its identifier. This operation is typically available only for admins.",
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the playlist to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Playlist with the given ID was not found.",
            "content": {

            }
          },
          "204": {
            "description": "Playlist successfully deleted."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "validationErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "TrackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "soundcloudUrl": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "artworkUrl": {
            "type": "string"
          }
        },
        "required": [
          "durationSeconds",
          "soundcloudUrl",
          "title"
        ]
      },
...
      "UrlRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "url"
        ]
      },
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "soundcloudUrl": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "totalTracks": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "soundcloudUrl",
          "title"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}